+++ cmd/repo/checks/runner.go
@@ -45,7 +45,7 @@ func (r *runner) run(ctx context.Context, arg string) error {
 		return microerror.Mask(err)
 	}
 
-	client, err := engine.GitHubClient(r.logger, r.flag.GithubTokenEnvVar, false)
+	client, err := engine.GitHubClient(r.logger, r.flag.GithubTokenEnvVar, false, nil)
 	if err != nil {
 		return microerror.Mask(err)
 	}